fastapi-app/requirements.txt
fastapi
uvicorn
python-multipart
opencv-python
tensorflow
numpy
pandas
dlib
sentence-transformers
faiss-cpu
scikit-learn
pydantic
